Move l2 precompiles logic to new module
