+++ crates/vm/levm/src/lib.rs
@@ -15,3 +15,4 @@ pub mod tracing;
 pub mod utils;
 pub mod vm;
 pub use environment::*;
+pub mod l2_precompiles;
