+++ crates/vm/levm/tests/tests.rs
@@ -2,9 +2,9 @@
 #![allow(clippy::unwrap_used)]
 
 use bytes::Bytes;
-use ethrex_levm::precompiles::bls12_pairing_check;
 #[cfg(feature = "l2")]
-use ethrex_levm::precompiles::p_256_verify;
+use ethrex_levm::l2_precompiles::p_256_verify;
+use ethrex_levm::precompiles::bls12_pairing_check;
 
 #[test]
 fn pairing_infinity() {
